Move command line functions back to command_line_utils.py
